feat: clean up stdout and document rationale
Previously the script would output errors about tools that were unavailable. Now the script doesn't
output errors for tools that may not have been installed by the script. The script also contains
documentation about how to silence these errors and a suggestion to note the tool was run but did
not produce any output, and why.