feat: gracefully handle the absence of nvidia-smi
Previously the script would produce errors if nvidia-smi wasn't avaialable, e.g., on a CPU node.
Now the script will direct nvidia-smi stderr to the same output file as stdout, and if the output
file is empty, will note in the ouptut file that nvidia-smi may not be available on the machine.