feat: streamline tool install checks
Previously the script required editors to manual call a function to check for and install tools.
The script now provides a dedicated indexed array to provide information about tools used for
information collection and tests. A single function call iterates through the array and performs
checks and installs as applicable. The script also outputs which packages may be installed for user
awareness, eliminating a hard-coded list that needed to be updated manually.