feat: create executable check
Previously the script suppressed output when a tool was not available, without storing information
about missing tools anywhere. This adds metadata for the tool executable check (pass/fail) which can
be accessed later on along with a function to update the tool metadata easily. Logic for the install
checks has been adjusted to ensure this happens regardless of whether tool installation is needed.