feat: add integer check and apply to update_needed_tools
Previously the script did not perform any santity checking to ensure two provided values were
integers. Now the script will exit if they are not and will instruct the user to report the issue to
Lambda support for further review. These errors would happen if the function were called incorrectly
by someone who edited the script and should not occur because of anything during runtime.