feat: validate the tools array makes sense
Previously the script did no checking and invalid information could be provided in the tools array.
The script now checks two of the metadata fields to ensure they are integers and exits with a
warning in the event they are not. This should only happen if the array is updated with invalid
information and should not otherwise occur during runtime.